refactor(agent): one runtime-ownership predicate behind two null policies
runtimeStatusAcceptsControl (messages/helpers.ts, fail-open) and
prompt-stream's runtimeIsActiveForPiSession / runtimeCanHydrateCanonicalSession
(both fail-closed) each hand-rolled the same 'does this runtime claim my pi
session' match. The match rule now lives once in runtime/session-status.ts as
runtimeClaimsPiSession, with the two policy wrappers keeping their deliberately
different missing-status handling — truth tables spelled out in doc comments.
runtimeIsActiveForPiSession's caller folds into runtimeCanHydrateCanonicalSession,
whose truth tables were identical over reachable inputs (every hydration caller
guards against an empty piSessionId before probing).
