+++ frontend/src/features/agent/runtime/prompt-stream.ts
@@ -16,10 +16,12 @@ import type {
 } from "@/features/agent/contracts";
 import type { BrowserBackend, ToolSelection } from "@/features/agent/tools/types";
 import * as api from "@/features/agent/runtime/api";
-import type { RuntimeStatus } from "@/features/agent/runtime/api";
 import { sessionRuntimeController } from "@/features/agent/runtime/session-runtime-controller";
 import type { Session, SessionId, UpdateSession } from "@/features/agent/runtime/types";
-import { settleTurn } from "@/features/agent/runtime/session-status";
+import {
+  runtimeCanHydrateCanonicalSession,
+  settleTurn,
+} from "@/features/agent/runtime/session-status";
 
 const EMPTY_SKILLS: ComposerSkillRef[] = [];
 const EMPTY_PROMPT_TEMPLATES: ComposerPromptTemplateRef[] = [];
@@ -184,7 +186,7 @@ function startPromptCommand(
           try: () => api.loadRuntimeStatus(context.runtime, currentPiSessionId),
           catch: () => null,
         });
-        if (runtimeIsActiveForPiSession(status, currentPiSessionId)) {
+        if (runtimeCanHydrateCanonicalSession(status, currentPiSessionId)) {
           deps.updateSession(context.sessionId, (session) => ({
             ...session,
             piSessionId: status?.piSessionId || session.piSessionId,
@@ -270,25 +272,4 @@ function mergeSkills(
   for (const skill of existing ?? []) byId.set(skill.id || skill.path || skill.name, skill);
   for (const skill of next) byId.set(skill.id || skill.path || skill.name, skill);
   return [...byId.values()];
-}
-
-function runtimeIsActiveForPiSession(
-  runtimeStatus: RuntimeStatus | null | undefined,
-  piSessionId: string | null | undefined,
-): boolean {
-  return Boolean(
-    runtimeStatus &&
-    runtimeStatus.active === true &&
-    (!runtimeStatus.piSessionId || !piSessionId || runtimeStatus.piSessionId === piSessionId),
-  );
-}
-
-export function runtimeCanHydrateCanonicalSession(
-  runtimeStatus: RuntimeStatus | null | undefined,
-  piSessionId: string,
-): boolean {
-  return Boolean(
-    runtimeStatus?.active === true &&
-    (!runtimeStatus.piSessionId || runtimeStatus.piSessionId === piSessionId),
-  );
 }
